+++ HCf460/application/matlab/time_detection.h
@@ -16,10 +16,10 @@
 
 /* Function Declarations */
 extern int time_detection(FIFO_DataType data[], int data_size, int win_size_time, int
-  stride_time, float time_times, int time_add);
+  stride_time, float time_times, float time_add);
 
 extern int quick_time_detection(FIFO_DataType data[], int data_size, int win_size_time, int
-  stride_time, double time_times, int time_add);
+  stride_time, float time_times, float time_add);
 
 #endif
